+++ src/main/java/com/buuz135/industrial/proxy/client/event/IFTooltipEvent.java
@@ -4,7 +4,7 @@ import com.buuz135.industrial.tile.block.IHasAdvancedTooltip;
 import net.minecraft.block.Block;
 import net.minecraft.client.gui.GuiScreen;
 import net.minecraft.init.Blocks;
-import net.minecraft.util.text.TextFormatting;
+import net.minecraft.util.text.TextComponentTranslation;
 import net.minecraftforge.event.entity.player.ItemTooltipEvent;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.relauncher.Side;
@@ -19,7 +19,7 @@ public class IFTooltipEvent {
         Block block = Block.getBlockFromItem(event.getItemStack().getItem());
         if (block instanceof IHasAdvancedTooltip && !((IHasAdvancedTooltip) block).getTooltip(event.getItemStack()).isEmpty()) {
             if (!GuiScreen.isShiftKeyDown()) {
-                event.getToolTip().add("Hold " + TextFormatting.YELLOW + TextFormatting.ITALIC + "Shift" + TextFormatting.RESET + TextFormatting.GRAY + " for Details");
+                event.getToolTip().add(new TextComponentTranslation("text.tooltip.hold_shift").getFormattedText());
             } else {
                 event.getToolTip().addAll(((IHasAdvancedTooltip) block).getTooltip(event.getItemStack()));
             }