+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 + Added a safety check for entities in the Mob Imprisonment Tool
 + Improved Log rendering
 + Fixed Item Splitter Facing
++ Tooltips now are translatable
++ Fixed Block Breaker and Block Placer accepting range addons
 
 # 1.6.5 
 + Added the mob in the Mob Imprisonment Tool name for easy searching